Add files via upload
Program that opens and reads a delimited data file that is located relative to the NetBeans project root folder. The delimited data file contains information about (in this example situation) prize winning music albums ranked by sales. The data file is called albums.txt. The data file should not be altered and should be considered as a read-only data file.